Made a mixin out of auto scroll utils
